Create a TopicCache
So that duplicate topic notification emails aren't repeatedly sent, the
TopicCache serves as an interface to a simple JSON document describing
the topics that have been sent.

Once an email is sent the corresponding topic is added to the cache. At
the end of all processing the cache is flushed to disk (as JSON).

TopicCache implements a dict-like API so it's easy to do `in`-type
membership testing of topics.