Fix authentication to private topics
Topics feeds can be accessed with multiple URLs:

1. community.lsst.org/t/topic-slug.json
2. community.lsst.org/t/topic-slug/topic-id.json
3. community.lsst.org/t/topic-id.json

Previously I used #1, but this was incompatible with Discourse auth;
the URL was being redircted to #2 and losing the auth params. So instead
we use the simplest form that doesn't redirect, #3.

TopicFeed still needs the slug to supply the HTML URL. Thus
CategoryFeed.new_topics now supplies a list of named tuples that
includes attributes with both the ID and slug of a topic.