Catch exception in link cleaning
With the post
members-of-lsst-dm-now-have-the-full-discourse-feature-set-trust-level-4
we ran into a weird issue with the link cleaner in contentpipe where a
link did not have an href field. So we're just going to catch the
KeyError exception and not process the link.

Next step is to find out how these pathological cases arise.